changed double to int

+++ include/globals.hpp
@@ -2,6 +2,6 @@
 #define CPP_TRIANGLES_INTERSECTION_GLOBALS_HPP
 
 constexpr double max_error = 1e-7;
-constexpr double triangle_points_num = 3;
+constexpr int triangle_points_num = 3;
 
 #endif
